# API/src/main/resources/static/swagger.yml
openapi: 3.0.0
info:
  title: PIR RDF plugin project API
  description: API documentation for the PIR RDF plugin project.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/validate:
    get:
      tags:
        - Auth
      summary: Validates a user account
      operationId: validateAccount
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account validated successfully
          content:
            text/plain:
              schema:
                type: string
                example: Account validated.
        "400":
          description: Bad request
          content:
            text/plain:
              schema:
                type: string
                example: Verification code expired. | Incorrect verification code.
        "404":
          description: User not found
          content:
            text/plain:
              schema:
                type: string
                example: Incorrect email.
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
  /api/auth:
    post:
      tags:
        - Auth
      summary: Create a new user account
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
              required:
                - email
                - password
                - firstName
                - lastName
      responses:
        "201":
          description: User created successfully
          content:
            text/plain:
              schema:
                type: string
                example: Verification code sent to email.
        "400":
          description: Bad request - email issue or other validation errors
          content:
            text/plain:
              schema:
                type: string
                example: Not possible to send email to this email address. | Incorrect email.
        "409":
          description: Conflict - duplicate key
          content:
            text/plain:
              schema:
                type: string
                example: Duplicate key email. | Data integrity violation.
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Mail dispatch service currently offline. | Oops, error. Please contact an administrator.
    get:
      tags:
        - Auth
      summary: Login to user account
      operationId: login
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Login successful
          content:
            text/plain:
              schema:
                type: string
                example:
                  {
                    "firstName": "<firstName>",
                    "lastName": "<lastName>",
                    "role": "<role>",
                    "id": <id>,
                    "token": "<token>",
                  }
        "403":
          description: Forbidden - bad credentials or account issue
          content:
            text/plain:
              schema:
                type: string
                example: Bad credentials
        "404":
          description: User not found
          content:
            text/plain:
              schema:
                type: string
                example: Incorrect email.
        "400":
          description: Bad request
          content:
            text/plain:
              schema:
                type: string
                example: Account locked.
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
    patch:
      tags:
        - Auth
      summary: Request password reset
      operationId: passwordForgotten
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Password reset requested successfully
          content:
            text/plain:
              schema:
                type: string
                example: Verification code sent to email.
        "404":
          description: User not found
          content:
            text/plain:
              schema:
                type: string
                example: Incorrect email.
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
    put:
      tags:
        - Auth
      summary: Set new password
      operationId: newPassword
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Password changed successfully
          content:
            text/plain:
              schema:
                type: string
                example: Password changed.
        "404":
          description: User not found
          content:
            text/plain:
              schema:
                type: string
                example: Incorrect email.
        "500":
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: Oops, error. Please contact an administrator.
  /api/user:
    get:
      tags:
        - User
      summary: Get all users
      operationId: getUsers
      responses:
        "200":
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden - user not authorized
          content:
            text/plain:
              schema:
                type: string
                example: Permission denied. Disappear !
...